[ARM32] Fix for Kingdom Hearts Chain of Memories
Kingdom Hearts crashes on the intro seemingly on all versions - have traced this back to gpsp 0.8, where it was working.  Adding an ARM equivalent of the MIPS check_store_reg_pc_no_flags(rd) in the arm_access_memory_load function appears to fix it on my platform (no crash on the intro after adding this line and building)?  @davidgfnet can you test and confirm?