[ARM32] Block Memory timing tweak (fixes Kingdom Hearts)
Moving the cycle_count increment out of the loop and just making it a single increment for these operations fixes the crash in the video at the beginning of Kingdom Hearts described in issue #183 .  This crash is intermittent - it will usually happen when starting the game 'cold' (i.e. first game after loading up RetroArch), and happens about 4-6 times out of 10 starts of the game.  

This tweak to timing results in 10 successful starts of the game with no crash.

NOW - I don't know if this will have a negative effect on other games.  Anecdotally other games I've tried seem to work ok, but I guess a run with Miniretro would probably be the proof.  I made this change after reading some of Exophase's old notes on how timing tweaks are necessary to fix some games (Mario and Luigi SS Saga is another case in point).