Fix Noise Channel (Add Noise Index from TempGBA)
A user on RGH Discord group pointed out that certain sound effects were missing.  (Sonic Advance was used as an example, when jumping on a box or enemy).  After investigating with the user, we found that it was specifically noise sound effects on GBC Sound Channel 4, and actually a different sound would be triggered in its place on another channel!

I tested the ROM on TempGBA and found the sound was correctly emulated.  This appears to be down to a separate index being used for the noise channel, so I have implemented it in the code here.  I have tested Sonic Advance and it fixes the issue (i.e. a proper 'crunch' sound is now played when jumping on a box, which is correct behaviour).