Fix Missing Sound on GBC Channel 4
A user on RGH Discord group pointed out that certain sound effects were missing. (Sonic Advance was used as an example, when jumping on a box or enemy). After investigating with the user, we found that it was specifically noise sound effects on GBC Sound Channel 4 that were missing.

I tested the ROM on TempGBA and found the sound was correctly emulated. I have therefore implemented the differences for noise sample handling from TempGBA to this fork of gpsp in this PR. I have tested Sonic Advance and it fixes the issue (i.e. a proper 'crunch' sound is now played when jumping on a box, which is correct behaviour).

The previous PR I created for this made more changes than necessary - as it turns out, just the changes in this PR are needed to fix the issue.