Banjo Kazooie GR - Intro black screen fix
This fixes the black screen on the intro for Banjo Kazooie.  Fix is taken from gpsp Kai.  In gpsp Kai this is a separate function used only by bitmap_render_pixel_mode4, so it may need to be here as well, but testing the combined function so far with other games doesn't appear to cause any issues.