Fix OBJ/Bitmap Ordering on Scanline Render
The previous code simply rendered layers in order of position, regardless of type.  The correct behaviour is that OBJs are rendered first and then backgrounds, so I have changed both the standard and conditional functions to reflect this.  This fixes the menus in Payback and will possibly restore missing graphics in other games.