Add save_type to game_config.txt to allow Backup overrides
The 'save_type' field in game_config.txt doesn't currently work, due to no code being in place to process it.  It should provide an override in the case where the backup method (e.g. Flash, SRAM or EEPROM) is incorrectly detected by gpsp and thus prevents the game from being able to save.  This also causes the copy protection to trigger in some games, such as the Dragon Ball series.

We should also add this code to the processing of gba_over.h, so that entries in there can be processed in the same way.  The good thing about game_config.txt is that it is user-editable, so people can add game entries themselves and periodically we can merge with gba_over.h.

save_type can have 3 possible values - flash, sram or eeprom.  gpsp should do the rest in determining the necessary size.