Make load_backup call unconditional (proposal for libretro saves)
Suggest moving the libretro save method check into load_backup.  This allows us to still try and get the size from a pre-existing backup file even if libretro save method is enabled - but we use the condition to only load the file if we're using gpsp method (mirrors current behaviour).